refactor: move sending utils to a different directory
